refactor fabs_nightly_loader to accept submission ids amongst other things
