+++ usaspending_api/broker/helpers/get_business_categories.py
@@ -1,191 +1,4 @@
-from distutils.util import strtobool
-
-
-def build_legal_entity_booleans_dict(row):
-    le_cols = [
-        'small_business_competitive',
-        'city_local_government',
-        'county_local_government',
-        'inter_municipal_local_gove',
-        'local_government_owned',
-        'municipality_local_governm',
-        'school_district_local_gove',
-        'township_local_government',
-        'us_state_government',
-        'us_federal_government',
-        'federal_agency',
-        'federally_funded_research',
-        'us_tribal_government',
-        'foreign_government',
-        'community_developed_corpor',
-        'labor_surplus_area_firm',
-        'corporate_entity_not_tax_e',
-        'corporate_entity_tax_exemp',
-        'partnership_or_limited_lia',
-        'sole_proprietorship',
-        'small_agricultural_coopera',
-        'international_organization',
-        'us_government_entity',
-        'emerging_small_business',
-        'c8a_program_participant',
-        'sba_certified_8_a_joint_ve',
-        'dot_certified_disadvantage',
-        'self_certified_small_disad',
-        'historically_underutilized',
-        'small_disadvantaged_busine',
-        'the_ability_one_program',
-        'historically_black_college',
-        'c1862_land_grant_college',
-        'c1890_land_grant_college',
-        'c1994_land_grant_college',
-        'minority_institution',
-        'private_university_or_coll',
-        'school_of_forestry',
-        'state_controlled_instituti',
-        'tribal_college',
-        'veterinary_college',
-        'educational_institution',
-        'alaskan_native_servicing_i',
-        'community_development_corp',
-        'native_hawaiian_servicing',
-        'domestic_shelter',
-        'manufacturer_of_goods',
-        'hospital_flag',
-        'veterinary_hospital',
-        'hispanic_servicing_institu',
-        'foundation',
-        'woman_owned_business',
-        'minority_owned_business',
-        'women_owned_small_business',
-        'economically_disadvantaged',
-        'joint_venture_women_owned',
-        'joint_venture_economically',
-        'veteran_owned_business',
-        'service_disabled_veteran_o',
-        'contracts',
-        'grants',
-        'receives_contracts_and_gra',
-        'airport_authority',
-        'council_of_governments',
-        'housing_authorities_public',
-        'interstate_entity',
-        'planning_commission',
-        'port_authority',
-        'transit_authority',
-        'subchapter_s_corporation',
-        'limited_liability_corporat',
-        'foreign_owned_and_located',
-        'american_indian_owned_busi',
-        'alaskan_native_owned_corpo',
-        'indian_tribe_federally_rec',
-        'native_hawaiian_owned_busi',
-        'tribally_owned_business',
-        'asian_pacific_american_own',
-        'black_american_owned_busin',
-        'hispanic_american_owned_bu',
-        'native_american_owned_busi',
-        'subcontinent_asian_asian_i',
-        'other_minority_owned_busin',
-        'for_profit_organization',
-        'nonprofit_organization',
-        'other_not_for_profit_organ',
-        'us_local_government'
-    ]
-    return {le_col: bool(strtobool(row.get(le_col, 'false') or 'false')) for le_col in le_cols}
-
-
-def get_award_category(award_type_code):
-    if award_type_code in ('A', 'B', 'C', 'D'):
-        return 'contract'
-    if award_type_code in ('02', '03', '04', '05'):
-        return 'grant'
-    if award_type_code in ('06', '10'):
-        return 'direct payment'
-    if award_type_code in ('07', '08'):
-        return 'loans'
-    if award_type_code == '09':
-        return 'insurance'
-    if award_type_code == '11':
-        return 'other'
-    return None
-
-
-def get_assistance_type_description(type):
-    if type == '02':
-        return 'BLOCK GRANT'
-    if type == '03':
-        return 'FORMULA GRANT'
-    if type == '04':
-        return 'PROJECT GRANT'
-    if type == '05':
-        return 'COOPERATIVE AGREEMENT'
-    if type == '06':
-        return 'DIRECT PAYMENT FOR SPECIFIED USE'
-    if type == '07':
-        return 'DIRECT LOAN'
-    if type == '08':
-        return 'GUARANTEED/INSURED LOAN'
-    if type == '09':
-        return 'INSURANCE'
-    if type == '10':
-        return 'DIRECT PAYMENT WITH UNRESTRICTED USE'
-    if type == '11':
-        return 'OTHER FINANCIAL ASSISTANCE'
-    return 'UNKNOWN ASSISTANCE TYPE'
-
-
-def get_business_type_description(business_type):
-    if business_type in ('A', '00'):
-        return 'STATE GOVERNMENT'
-    if business_type in ('B', '01'):
-        return 'COUNTY GOVERNMENT'
-    if business_type in ('C', '02'):
-        return 'CITY OR TOWNSHIP GOVERNMENT'
-    if business_type in ('D', '04'):
-        return 'SPECIAL DISTRICT GOVERNMENT'
-    if business_type == 'E':
-        return 'REGIONAL ORGANIZATION'
-    if business_type == 'F':
-        return 'U.S. TERRITORY OR POSSESSION'
-    if business_type in ('G', '05'):
-        return 'INDEPENDENT SCHOOL DISTRICT'
-    if business_type in ('H', '06'):
-        return 'PUBLIC/STATE CONTROLLED INSTITUTION OF HIGHER EDUCATION'
-    if business_type == 'I':
-        return 'INDIAN/NATIVE AMERICAN TRIBAL GOVERNMENT (FEDERALLY RECOGNIZED)'
-    if business_type == 'J':
-        return 'INDIAN/NATIVE AMERICAN TRIBAL GOVERNMENT (OTHER THAN FEDERALLY RECOGNIZED)'
-    if business_type in ('K', '11'):
-        return 'INDIAN/NATIVE AMERICAN TRIBAL DESIGNATED ORGANIZATION'
-    if business_type == '12':
-        return 'OTHER NON-PROFIT'
-    if business_type == 'L':
-        return 'PUBLIC/INDIAN HOUSING AUTHORITY'
-    if business_type == 'M':
-        return 'NONPROFIT WITH 501C3 IRS STATUS (OTHER THAN INSTITUTION OF HIGHER EDUCATION)'
-    if business_type == 'N':
-        return 'NONPROFIT WITHOUT 501C3 IRS STATUS (OTHER THAN INSTITUTION OF HIGHER EDUCATION)'
-    if business_type in ('O', '20'):
-        return 'PRIVATE INSTITUTION OF HIGHER EDUCATION'
-    if business_type in ('P', '21'):
-        return 'INDIVIDUAL'
-    if business_type in ('Q', '22'):
-        return 'FOR-PROFIT ORGANIZATION (OTHER THAN SMALL BUSINESS)'
-    if business_type in ('R', '23'):
-        return 'SMALL BUSINESS'
-    if business_type == 'S':
-        return 'HISPANIC-SERVING INSTITUTION'
-    if business_type == 'T':
-        return 'HISTORICALLY BLACK COLLEGES AND UNIVERSITIES (HBCUS)'
-    if business_type == 'U':
-        return 'TRIBALLY CONTROLLED COLLEGES AND UNIVERSITIES (TCCUS)'
-    if business_type == 'V':
-        return 'ALASKA NATIVE AND NATIVE HAWAIIAN SERVING INSTITUTIONS'
-    if business_type == 'W':
-        return 'NON-DOMESTIC (NON-US) ENTITY'
-    if business_type in ('X', '25'):
-        return 'OTHER'
-    return 'UNKNOWN TYPES'
+from usaspending_api.broker.helpers.build_legal_entity_booleans_dict import build_legal_entity_booleans_dict
 
 
 def get_business_categories(row, data_type):
@@ -531,10 +344,3 @@ def get_business_categories(row, data_type):
     else:
         raise ValueError("Invalid object type provided to update_business_categories. "
                          "Must be one of the following types: TransactionFPDS, TransactionFABS")
-
-
-def set_legal_entity_boolean_fields(row):
-    """ in place updates to specific fields to be mapped as booleans """
-    legal_entity_bool_dict = build_legal_entity_booleans_dict(row)
-    for key in legal_entity_bool_dict:
-        row[key] = legal_entity_bool_dict[key]
